hotels_telegram_bot/requirements.txt
requests~=2.31.0
peewee~=3.16.3
python-dotenv~=1.0.0
pydantic~=2.3.0
loguru~=0.7.1
telebot~=0.0.5
django-scheduler~=0.10.1